requirements.txt
python-bitbucket
boto3
docker
GitPython